testsuite/helm-chart-validator/requirements.txt
pytest==6.1.2
pytest-html==3.0.0
pyyaml==5.3.1
semver==2.13.0
memoization==0.3.1
